+++ GayHaikuTabbed/GHComposeViewController.h
@@ -25,8 +25,8 @@
     UITextView *textView;               //Editable text view for haiku entry.
     UITextView *nextInstructions;       //Contains the instructions for moving from instructions to compose (should be a label?)
     UIAlertView *alert;                 //Alert in case of syllable errors.
-    float screenHeight;
-    float screenWidth;
+    int screenHeight;
+    int screenWidth;
     CGRect frame;
     GHHaiku *ghhaiku;                   //Instantiation of GHHaiku.
     GHVerify *ghverify;                 //Instantiation of GHVerify.
